[dawn] Fix modernize-avoid-c-style-cast, pt 2
- Non-trivial replacements for T(x)

Bug: 548071770
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/334539
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/dawn/native/Buffer.cpp
@@ -84,7 +84,7 @@ class ErrorBuffer final : public BufferBase {
         DAWN_CHECK(!mFakeMappedData);
 
         uint64_t size = GetSize();
-        if (size < uint64_t(std::numeric_limits<size_t>::max())) {
+        if (size < std::numeric_limits<size_t>::max()) {
             mFakeMappedData =
                 // SAFETY: Frontend is responsible for initializing MapAtCreation memory.
                 DAWN_UNSAFE_BUFFERS(
