+++ src/dawn/native/vulkan/RenderPipelineVk.cpp
@@ -521,9 +521,9 @@ ResultOrError<RenderPipeline::SpecializationResult> RenderPipeline::InitializeSp
     std::optional<uint32_t> pixelCenterPolyfillLocation = std::nullopt;
     if (NeedsPixelCenterPolyfill()) {
         const EntryPointMetadata* vtx = GetStage(SingleShaderStage::Vertex).metadata;
-        for (size_t i = 0; i < vtx->usedInterStageVariables.size(); ++i) {
+        for (uint32_t i = 0; i < vtx->usedInterStageVariables.size(); ++i) {
             if (vtx->usedInterStageVariables[i] == false) {
-                pixelCenterPolyfillLocation = uint32_t(i);
+                pixelCenterPolyfillLocation = i;
                 break;
             }
         }
